PublishedProject.zip_name: add legacy argument.
We would like to avoid including the project title in zip file names
as well as in the zip file contents, in favor of the project slug,
which is more stable and machine-readable.

Eventually, therefore, the behavior of the zip_name function should be
changed; as a first step in that direction, add an optional argument
so callers can specify whether they want the old or new style.
Existing code continues to use the old style for now.