console_navbar: fix lots of broken and inconsistent IDs.
HTML IDs must be unique.  Furthermore, to avoid confusion or possible
clashes, all of the IDs used in the navbar should begin with "nav_".

All of the #sectionComponents are renamed to #nav_section_components
(corresponding to #nav_section_dropdown), for consistency.

The "credentialing" section is renamed to "identity".  The "usage"
section is renamed to "stats".

For consistency, ensure all of the page links have an ID.  All of
these links are renamed to #nav_viewname (corresponding to
{% url 'viewname' %}.)

For those links that kludgily point to a URL with a mandatory
parameter, use #nav_viewname__arg.