console_home: require the user.can_view_admin_console permission.
This view itself is somewhat of a placeholder and doesn't currently
show any important information.  But it also doesn't hurt to add a
permission requirement so that we don't have to treat this view as a
special case.

The user.can_view_admin_console permission controls whether the "Admin
Console" link appears in the nav bar, so it makes sense to restrict
this view to people who have the user.can_view_admin_console
permission.