console_navbar: replace hardcoded links with a generic template.
The new class NavMenu is used to construct the navigation menu links.

Previously, console_navbar.html determined which links and submenus
should be displayed by hardcoding the permissions, which needed to
match the corresponding permissions for the view.  This is now done
automatically, by relying on the @console_permission_required
decorator.

Previously, in order to determine which link should be highlighted as
"active" and which submenu should be visible by default, there was a
separate "something_nav" parameter for every menu item, and the
corresponding view needed to set the correct parameter by hand.  This
is now done automatically by comparing the request path to the link
path.

(This relies on the assumption that if page X is a logical
sub-component of page Y, then URL Y is a prefix of URL X.  That
assumption might be wrong in some cases - in which case we should
change the URLs to paths that make more sense, and perhaps add
redirections if needed.

Note in particular that every page in the console probably ought to
have exactly one of the menu-item URLs as a prefix.)