console.views: new view cloud_mirrors.
This page provides a list of published projects, and shows whether
each project is currently mirrored on one of the supported cloud
storage platforms.

This page currently just displays a read-only list, and doesn't offer
any way to change the state of cloud mirrors (that must be done
through the manage_published_project view.)

Note that to efficiently display the projects and their associated
cloud mirror information, we want to make a single query using
select_related().  In particular, we don't want to select all
PublishedProjects and then individually check each project's gcp and
aws attributes in a Python loop, which would perform 2*N+1 queries.

Furthermore, since the list of projects is likely to be large, we use
only() to avoid querying unnecessary columns.  This should be
carefully tested to be sure that the template (cloud_mirrors.html)
doesn't directly or indirectly reference unlisted columns.