project_submission: handle case of an unassigned project.
On the submission page we display a contact address for the project
editor, if one has been assigned.  If no editor has yet been assigned,
then the "contact_email" won't be shown.

However, if no editor has yet been assigned (self.editor is None) and
there is no PROJECT_EDITOR_EMAIL defined, the previous code would
crash.

Put this logic into the model class, not the view, and document it.