LinkFilter: permit relative subresource URLs even without base_url.
The LinkFilter class is used to sanitize and normalize URL attributes
in uploaded HTML content.  In the case of subresources (such as
images) we want to disallow any that point to an external server.

This was broken in the case of subresource URLs that were already
host-relative or document-relative, in the case where the document URL
(base_url) was unspecified.

For example, if the base_url was set to 'https://example.com/foo/' and
the input was '<img src="xyz.jpg">', the URL would be resolved to
'https://example.com/foo/xyz.jpg', which would then be normalized to
'/foo/xyz.jpg', which would be allowed.  But if the input was
'<img src="/foo/xyz.jpg">' and the base_url was None, this would be
rejected.

SafeHTMLField invokes LinkFilter with base_url set to None, because
SafeHTMLField itself has no idea what URL or URLs will be used for
publishing the content.  But using a host-relative or even a
document-relative URL should always be "safe".  (In some cases we
might *want* to disallow document-relative URLs, which is why we have
Metadata.update_internal_links(); but that shouldn't be a security
concern.)