physionet.settings.{production,staging}: use new config vars.
Instead of hard-coding the database hostname, username, and database
name, use the config variables DB_HOST, DB_USER, and DB_NAME.
Additionally, replace the old DATABASES_PASSWORD with DB_PASSWORD.
All four new variables are now required to be defined in .env.

This is intended to bring the old settings modules in line with
physionet.settings.settings, so that eventually all three can be
unified.