edit_certification: remove unused form processing logic.
The page at /settings/certification/ does not contain any forms.  As
far as I can tell, there are no forms on the site that post to
/settings/certification/.  (The place where people actually submit
training applications is /settings/training/, which posts to
/settings/training/.)

The POST-handling code here was apparently copied and pasted from
edit_training and has no purpose.