Make the "no_phi" field an "internal" quality assurance check.
When reviewing a project, the editor is asked to perform a series of
"quality assurance" checks.  In the "edit submission" form
(console.forms.EditSubmissionForm), they must answer "yes", "no", or
"undetermined" for each QA check.  All answers must be "yes" in order
to accept the project.

One of the questions is "Is the project free of protected health
information?"  We don't want to give anyone the false impression that
the editor has performed an exhaustive review or that this constitutes
an "expert determination".  In fact, the author is responsible for
doing that review themselves before *uploading* the data.

Therefore, keep this field among the questions included in
EditSubmissionForm, but do not list it as a "quality assurance result"
that is shown to the authors.  Note that the public "quality assurance
results" are generated by EditLog.set_quality_assurance_results, and
they are included in email messages:
- notification/templates/notification/email/accept_submission_notify.html
- notification/templates/notification/email/reject_submission_notify.html
- notification/templates/notification/email/revise_submission_notify.html
and in the submission timeline:
- project/templates/project/active_submission_timeline.html
- project/templates/project/static_submission_timeline.html