project/urls.py: add parameters for TestURLs.
Several views in this app have mandatory query parameters - these
views are accessed only by JavaScript, and return an HTML
fragment rather than a complete page.

A couple of views are not yet tested because the demo database
does not contain any UploadedDocument objects or any DataAccess
objects.  This should be fixed.