generate_doi_payload: use publication date, not current date.
When creating or updating a DOI, the metadata must include the
publication year.

If the DOI is being published (or has already been published and is
being updated), then the project must already be published and the DOI
publication year should match the PublishedProject's publish_datetime.

If the DOI is a draft, then the project need not be published
already (it may still be an ActiveProject which has no
publish_datetime.)  Since it's a draft, use the current year as a
placeholder.