manage_published_project: disable inapplicable "core DOI" buttons.
In the management page for a published project are buttons to create
or update the DOI metadata, both for the "versioned" and "core" DOIs.

The "core" DOI points to the latest project version (i.e, it points to
'published_project_latest', which redirects to the single
PublishedProject that has is_latest_version=True.)  Therefore, the
metadata of this DOI (e.g., authors, abstract, publication year) needs
to reflect the latest project version.

Previously, the management page would allow you to erroneously
create/update a core DOI from a non-latest-version project.  Instead,
return an error message if you try to do this, and disable the buttons
to make it clear this is not allowed.