PublishedReference: add url field.
This field should hold the URL of the cited work, if any.

In most cases, this should be easy to determine by searching for a
regular expression.  However, if that pattern is changed in the
future, we may not want the new rules to apply retroactively.

Therefore, for published projects, store the URL in a model field, and
perform a one-time migration to compute this field for existing
published projects.