console.test_views: fix test cases for editor assignment.
The previous test case in test_assign_editor did not check that the
project was assigned to the correct editor (it asserted that
project.editor was true, not that project.editor equalled editor.)

The previous test case in test_reassign_editor likewise did not check
that the project was correctly reassigned.

And indeed, test_reassign_editor was *not* correctly reassigning the
project, because it did not include the required 'reassign_editor'
parameter.

Moreover, test_reassign_editor was not logging in as the assigned
editor before attempting reassignment - in fact this currently works,
but it isn't supposed to be allowed.

Additionally, rearrange the code a little for readability, use
refresh_from_db instead of duplicating earlier queries, and add
additional precondition checks.