console.test_views: test creating DOIs for a published project.
Test that the manage_published_project page can be used to create new
DOIs for a published project that doesn't have any.  In this case, the
DOI API server is asked to assign unused (unique, randomly-generated)
DOIs, and also to make the new DOIs public immediately.

The publication year for both new DOIs should equal the project's
publication year (not the current year.)