console.test_views: test updating DOIs for a published project.
Test that the manage_published_project page can be used to update DOIs
for a published project that already has them.  In this case, the DOI
API server is asked to modify the existing DOIs, not to assign new
ones.

The publication year for both DOIs should be set to the project's
publication year (not the current year), and both DOIs should be made
public if they weren't already.