NewProjectVersionForm: set order field for references.
Reference (and PublishedReference) objects have an "order" field which
is supposed to determine the order the references are displayed.

This field is not set or used consistently.  Try to make the situation
less bad for new active projects, by creating references in order (so
that the order of their "id" will match the order of their "order")
and also copy the "order" from the published project.