Fix Principal value in verification bucket policy.
The Principal listed in a bucket policy needs to follow the correct
syntax or it will be rejected as a bad bucket policy.

configure_aws_verification_bucket is used to set the policy for the
user verification bucket.  I have no idea how I committed this with a
broken policy but it seems like this never worked to begin with :)

There shouldn't be any security implications to allowing any principal
to access the bucket.  We're verifying on the server side that the
appropriate parameters are included in the URL, which implies that the
requester is actually logged in to an AWS user account.