ReferenceFormSet: Allow fixing invalid reference order.
If a project has an invalid order of references, we must provide a way
for the author or editor to review and fix the order.

The past behavior of the "project content" and "copyedit" pages was
that ReferenceFormSet would set the 'order' field for every form in
the formset.  The effect of that, however, was that 'order' would be
set for:
  (a) newly added references
  (b) existing references whose 'description' was changed
References that already existed, and whose 'description' field was
unchanged, would not have been saved (because their form.changed_data
was False.)

Since many existing projects may have scrambled lists of references
and can't be fixed automatically, we want broken projects to be
"flagged", and we want that flag to persist until someone is able to
review the project and check that it's fixed.  In particular, simply
opening the "project content" page and submitting it, without making
any changes, should not cause the flag to be cleared.

Therefore: if the project has an invalid reference order, then display
a warning message with a checkbox (at the bottom of the page, below
the reference list and above the submit button.)

By default, that checkbox isn't checked, and in that situation,
ReferenceFormSet should not alter any existing reference 'order'.
Newly-added references should have a correct 'order', but if existing
references were "invalid", they should remain "invalid".

If the checkbox is checked, then ReferenceFormSet should set all
reference 'order' to match the current form order (which is what the
old code was trying to do, and didn't work because it didn't set
changed_data.)