Try to unscramble references in simple cases.
Quite some time ago an "order" field was added to the Reference and
PublishedReference models, to make the order explicit.

However, this field was neither required to be non-null, nor required
to be unique, so nothing prevented creating and publishing a project
with an ill-defined reference order.  In particular:

(a) NewProjectVersionForm did not copy "order" from PublishedReference
    into Reference; so newly created versions would have unpredictable
    ordering.

    (This bug has been fixed by commit 6b4261365.)

(b) ReferenceFormSet tried to set "order", but only succeeded in saving
    the value for references that were being created/modified; so if a
    reference was deleted and then new references added, there could
    be multiple references with the same "order" value.

    (This bug is still present.)

It is difficult, bordering on impossible - especially after a project
has been published - to guess the author's intent and fix the
resulting issues automatically.

However, a common situation is that two versions of the same project
have *identical* lists of references, and the second version has all
"order" values set to None.  That means that the author created a new
version of the existing project and *never edited* any of the
references; so it is probably safe to assume that the author intended
to keep the order exactly as it was in the previous version.

An example of this would be:

- https://web.archive.org/web/20240418010540/https://physionet.org/content/annotation-dataset-sdoh/1.0.0/
- https://web.archive.org/web/20240710153623/https://physionet.org/content/annotation-dataset-sdoh/1.0.1/

A slightly less common situation is that the second version of the
project has all the same references as the first version, but also has
one or more additional references added to (what the author intended
to be) the end of the list.  Since those new references would have
been created via ReferenceFormSet, the new references have "order"
greater than the number of old references, while the old references
still have "order" equal to None.

Again, if the old references were never edited, it's probably
reasonable to assume that the author didn't intend to change their
order.

An example of this would be:

- https://web.archive.org/web/20240731053747/https://physionet.org/content/mimiciv/2.2/
- https://web.archive.org/web/20240806233833/https://physionet.org/content/mimiciv/3.0/

(notice that the three new references are listed as the *first* three
items, because null is sorted after any non-null value.)

Here, we add a data migration that should automatically fix those two
cases, and other similarly unambiguous cases.  It works by iterating
over projects in publication order, meaning that it should be able to
repair references that were copied from version A to version B, and
then later copied from version B to version C.

One thing worth noting here is that we do not simply change the
"order" of existing Reference and PublishedReference objects.
ReferenceFormSet has never honored the "order" when it comes to
displaying references in the Project Content (or Copyedit) pages.
That formset has always ordered references by their primary key
("id").  That logic would be difficult to change, and doing so would
only make the problem worse by further obscuring the author's intent.
On the other hand, it also seems impractical to revert to
"id"-ordering in previews and published projects.

Thus, when repairing a broken project, we actually shuffle the
"descriptions" of the existing Reference objects, so that the final
"order" and "id" should result in the same actual ordering.  It's
possible that this could lead to errors if an author is trying to edit
the reference list while this migration is being deployed, but that
seems fairly unlikely and less bad than the alternatives.  It's not
really necessary to do this for PublishedReferences, but using the
same logic as for References doesn't hurt.