schema_metadata.json: do not escape arguments to url tag.
{%url%} is not a JSON or JavaScript parser; it makes no sense to apply
JavaScript escaping to its arguments.  If the slug is "x-y", then we
want to link to /content/x-y/1.0.0/, not /content/x%5Cu002Dy/1.0.0/.

If the *output* of {%url%} needed to be escaped, then it would be
appropriate to embed the {%url%} in a {%filter%}.  But in fact,
{%url%} already escapes " as %22, so this would be a non-issue even if
it were possible to use " in project slug or version, which it isn't.