s3: new functions for querying existing objects.
get_s3_object_info() retrieves the metadata of a single existing
object in an S3 bucket.  list_s3_subdir_object_info() retrieves
metadata of all objects in a subdirectory-style prefix.

S3ObjectInfo contains the metadata that is needed in order to check
whether two S3 objects match, or to check whether an S3 object matches
a local file.

S3 has a complicated way of calculating the ETag.  If the object is
uploaded in a single request then the ETag is the MD5 hash of the
content, but if it's a multi-part upload then the ETag is a hash of
the hashes of the parts.

boto3, by default, splits files into 8 MiB chunks (and for files of
exactly 8 MiB, it does a "multi-part" upload with exactly one part);
but there is no guarantee that other clients will do the same.  It's
infeasible to verify ETag in the general case without knowing the
algorithm used by the uploading client.  However, for our purposes, it
should be okay to assume that the file was uploaded by boto3 using the
default settings.