test_s3: test re-uploading a project after modifying its content.
upload_project_to_S3() should now check to see which files already
exist in the destination bucket, and avoid re-uploading those that
haven't changed.

We want to test that this is the case, which we can do by adding
custom "object tagging": when an existing object is replaced by
re-uploading, any tags on the existing object are erased.

We also want to test that this works correctly for objects of 8 MiB
and larger, because the ETag calculation is not trivial (see
project.cloud.s3.s3_file_etag).