project.test_views: test handling of invalid reference order.
Past bugs in the project editing forms have allowed project references to
be created with 'order=None', or with two references having the same
'order'.  Either situation should now prevent the project from being
submitted; check that this is the case.

ReferenceFormSet should allow these situations to be fixed manually:
if the confirm_reference_order checkbox is checked, then the 'order'
of all references should be updated; and if that box is not checked,
the 'order' of existing references should be untouched.  Check that
this works as expected.