ActiveProject: ethics_statement must be non-blank for submission.
The ethics_statement field is required, and a project should not be
submitted or published if the field is blank.

Previously, ethics_statement was not included in REQUIRED_FIELDS but
was handled separately.  However, the previous code permitted blank
HTML values (such as "<p>&nbsp;</p>"), which are not allowed for
fields in REQUIRED_FIELDS.

Additionally, REQUIRED_FIELDS is reformatted for better readability.