edit_cloud: add links to instructions for using AWS CLI.
We need to tell people that they need to be using an IAM user.

We also need to tell them explicitly to enable "S3 full access": if
your user doesn't have an IAM user policy that says they can access
S3, then that user cannot access S3 resources in other AWS accounts,
regardless of those resources' policies.  (This is weird and
surprising, because you can access S3 resources in your own account in
that case; it's also weird and surprising that the default IAM user is
forbidden from doing things that anonymous users are allowed to do.)