On the Project Submission Info page, when there is a new version of a published project, a message is displayed alongside the version number stating, 'This is an update of the latest published version. Otherwise, a message stating 'No latest version available' is displayed. #2282 closed.
On the Project Submission Info page, when there is a new version of a published project, a message is displayed alongside the version number stating, '(this is an update to a published project)'. Otherwise, a message stating 'Latest Published Version: No published version available.' is displayed. #2282 closed.

On the Project Submission Info page, when there is a new version of a published project, a message is displayed alongside the version number stating, '(this is an update to a published project)'. Otherwise, a message stating 'Latest Published Version: No published version available.' is displayed. #2282 closed.

On the Project Submission Info page, when there is a new version of a published project, a message is displayed alongside the version number stating, '(this is an update to a published project)'. Otherwise, a message stating 'Latest Published Version: No published version available.' is displayed. #2282 closed.