install-pn-test-server: allow unsafely cloning local directories.
Recent versions of git in Debian bookworm have broken the ability to
clone a local directory that is owned by another user.  (It's not
obvious to me whether this is a long-term problem in git upstream or a
temporary problem in Debian.)

install-pn-test-server fully trusts the selected source repositories -
i.e., we intend on executing code from them - so marking these as
"safe" is an annoyance but should not have security impact.