published_project: hide "total uncompressed size" if unknown.
If the total download size for a project is not (yet) known, the
PublishedProject's main_storage_size is zero.  (Currently, this value
is not known until the move_files_as_readonly task has finished; the
download size is obviously dependent on the size of SHA256SUMS.txt.)

In the event that the size is unknown, it should be hidden rather than
displaying "0 B".