project_files: indicate the maximum number of files that can be uploaded.
The file upload form is limited in the number of files it can handle
due to design limitations of the Django upload handler.  Exceeding the
limit set by DATA_UPLOAD_MAX_NUMBER_FILES will currently cause the
page to crash (and doesn't give a useful error message.)

Display a message to advise authors of this limitation.