project_files: disable form submission if too many files selected.
The file upload form is limited in the number of files it can handle
due to design limitations of the Django upload handler.  Exceeding the
limit set by DATA_UPLOAD_MAX_NUMBER_FILES will currently cause the
page to crash (and doesn't give a useful error message.)

To prevent this happening accidentally, disable the submit button and
show an error message if too many files are selected.