Add content_sections() and content_section_headers().
Project descriptions consist of many "sections" of free-form HTML
content, but the order of the sections, as well as which ones are
required/allowed, depends on the project type.  Add new methods
Metadata.content_sections() and ProjectType.content_section_headers()
that can be used to obtain the list of sections in a generic way.

The ContentSectionHeaders are hardcoded in projecttype.py.  The order
of sections, their titles, HTML IDs, and corresponding field names,
should match the existing template files:
 - database_content.html
 - software_content.html
 - challenge_content.html
 - model_content.html

A ContentSectionHeader is "required" if the field is listed in
ActiveProject.REQUIRED_FIELDS.