pnmath: use DOMParser to parse MathML content.
In TinyMCE 7, the tinymce.html.DomParser parses MathML elements as
opaque elements - the resulting 'math' node has no children (instead,
the MathML markup is stored as the 'value' of the node.)

We need to extract the text content of the annotation element, and it
doesn't appear that there's a way to do so using the TinyMCE
DomParser.  So, instead, use a DOMParser to do this.  It would
generally be unsafe to apply a DOMParser to untrusted content (that's
why tinymce.html.DomParser exists), but as far as I know it should be
safe if the untrusted nodes aren't inserted into any document [1].
Also note that the text we're passing to the DOM parser has, in
theory, already gone through TinyMCE's sanitization.

[1] https://stackoverflow.com/a/64772926