CoreProject: archived projects aren't "being worked on".
Authors are not allowed to create two active versions of the same core
project at a time; if a project has an active version, the "New
Version" button should not be shown, and new_project_version should
reject the request.

Whether an author is allowed to create a new version is determined by
PublishedProject.can_publish_new, which invokes
CoreProject.active_new_version to determine whether there's an
existing ActiveProject.

Originally, when an active project was "deleted" or rejected, the
ActiveProject object was deleted and an ArchivedProject object created
to replace it.  So afterwards, the author could go back and create a
new active version from scratch.

However, archived projects are now a subcategory of ActiveProjects.
Since they should not be counted as "being worked on", a
deleted/rejected submission should not prevent the author from
creating a new submission.