Replace ContentForm.FIELDS with content_sections().
ContentForm is the form used for editing (most of) the project
description content, used by the authors (project_content) and
editors (copyedit_submission).  This form has a dynamically generated
list of fields, which depend on the resource type.

Instead of embedding the list of fields here, use
instance.content_sections() to obtain the list of content fields in
display order.  The results should be identical.

(Note that ContentForm includes the 'title' field in addition to the
content fields, and it excludes 'ethics_statement' because that is
handled by EthicsForm instead.)