Replace hardcoded HTML templates with content_sections().
published_project.html defines the layout of the published project
page.  The main content of this page varies depending on the project
type, and thus the template is split into four separate files.

Likewise, project_preview.html defines the layout of the project
preview page (which should closely mirror published_project.html), and
its main content is split into four separate files.

Instead, these can be handled generically by calling
project.content_sections(), and removing the hardcoded structure.

In the case of the published project, any sections that are blank are
omitted, even if they are "required" (meaning that the project would
not be allowed to be published today.)  This generalizes the handling
of the ethics_statement field.

(For example, the demoecg project has a blank "Background" field, so
this empty section is now omitted.  In the real world, the Background
field has always been required, so the demoecg project would not have
been allowed to be published.)