Replace ActiveProject.REQUIRED_FIELDS with content_sections().
Many project fields are required for publication (for example, the
'abstract' field is allowed to be blank when the project is saved, but
it must be non-blank for the project to be submitted and published.)
The list of required fields is dependent on the project type.

In the case of HTML content section fields, each ContentSection has a
'required' attribute to indicate if it is required, and thus this
question can be answered generically via self.content_sections().

The remainder of the required fields (title, version, license,
short_description) are not project-type-dependent.