Replace ActiveProject.LABELS with content_sections().
The "label" used for each form field in ContentForm is the same as the
"title" of the section shown on the published project page.  In some
cases, the title isn't the same as the name of the model field, and
this depends on the project type.

The title can now be obtained generically via
instance.content_sections(), and since this means ActiveProject.LABELS
is no longer used, ActiveProject.LABELS can be removed.