ContentForm: get resource type from the project object.
ContentForm is the form used for editing (most of) the project
description content, used by the authors (project_content) and
editors (copyedit_submission).  This form has a dynamically generated
list of fields, which depend on the resource type.

This form is always created with an existing project instance, and the
form's "resource type" should always match the type of that project.
So instead of requiring an explicit resource_type argument (which is a
resource type ID, not a ProjectType object), take this information
from the project itself.

(Additionally, replace the use of old-style super with the equivalent
Python 3 syntax.)