published_project.html: always use serve_published_project_zip.
If a project has a zip archive that can be downloaded, we show a link
in the Files section.  Previously (for efficiency), we would display a
direct link to the static file if the project was public, or a link to
the serve_published_project_zip view if not.

However, the serve_published_project_zip view works for all types of
projects, is only marginally less efficient, and lets us consolidate
authorization logic, so use that in all cases.