lightwave_server: use per-project sandbox for published data.
The LightWAVE server runs in a sandbox to ensure that it can't be used
to access confidential data, and to guard against security bugs in
libwfdb, libFLAC, or the LightWAVE server itself.

Historically there were a number of published databases on PhysioNet
that included symbolic links to files in another database, which meant
that the sandbox chroot directory needed to include all of the
published databases at once.  This has now been addressed - all of the
symbolic links within the pn-static/published-projects directory on
the PhysioNet server have now been replaced with hard links or deep
directory copies - so it should be safe to restrict the chroot
environment according to the selected 'db'.  This will eventually
allow us to move open projects into the pn-media storage or to support
multiple storage backends.

This change does not handle non-open-access projects; those remain
unsupported for now.

For efficiency, we still want to minimize database traffic, so limit
the PublishedProject query to the columns that are relevant (either
for constructing the 'dblist', or for calling can_view_project_files()
and file_root().)  This should speed up 'action=dblist' and hopefully
should not slow down other operations substantially.

When the client requests a 'db' that isn't a published project or that
they don't have permission to access, we generate a JSON response
directly rather than invoking sandboxed-lightwave.  This is a 200
response (not a 403 or 404) to facilitate handling in JavaScript.