lightwave: test number of queries for published project access.
lightwave.views.lightwave_server performs a database query either to
retrieve the list of all published WFDB databases or to check that the
selected database is accessible.

For efficiency, that query uses only() to select only the columns that
are needed by can_view_project_files() and project.file_root().  If in
the future those functions are changed, the 'only' statement may need
to be updated accordingly.  Add test logic to verify that we are not
inadvertently making more than one database query.