physionet-upgrade-test: use old version's .env.example.
In general, if new variables are added to the .env file, they should
have sensible, working defaults, so that the administrator isn't
required to manually edit the .env file before installing the new
code.

The purpose of the physionet-upgrade-test workflow is to ensure the
new version can be cleanly installed on a running server.  To do this,
we use the same .env file for running both the old and new test
suites.  (As opposed to, for example, using the old .env.example for
the old test suite and the new .env.example for the new test suite -
which would not be a realistic simulation of an upgrade!)

Previously, we used the new .env.example for testing both versions.
However, if we instead use the old .env.example, this should help to
catch mistakes where new variables are added without having a working
default.