TestS3: add test for uploading controlled projects.
Uploading a controlled project, such as demoeicu-2.0.0, should work in
just the same way as uploading an open project, except that it goes to
a different bucket with a different bucket policy.  Add a test case
to check that file uploads work correctly.

Note that upload_project_to_S3 currently doesn't explicitly set a
PublicAccessBlockConfiguration for controlled buckets (which means, in
practice, that the bucket will use AWS default settings, which are
maximally restrictive.)