style.yml: move style checks to their own workflow.
Code style checks are independent of functional checks, and it makes
sense to handle them separately.  For example, it doesn't make sense
to run the same code style checks multiple times in different
environments.

The version of flake8 in Ubuntu 24.04 doesn't support --diff.  Rather
than using an older version of Ubuntu, or a Debian container as we do
in physionet-build-test.yml, just install the versions we want using
pip.