style.yml: simplify logic and remove useless "push" check.
When checking a pull request (i.e. a prospective merge commit), the
things we are comparing are HEAD (the merged result) and HEAD^ (the
current head of the destination branch.)  This will work both for
"pull_request" events and for "merge_group" events, and there's no
need to dig into GitHub API details.

Additionally, the purpose of this workflow is not to detect all style
errors in the codebase, just to detect style errors in modified code.
So there's not really any point in running this workflow "on push".
It's also not clear to me whether the "diff origin/dev HEAD" was ever
doing what it was intended to do.