workflows: don't run tests upon pushing to dev.
We now use merge queues ("merge_group" events) to test each commit
before it is pushed to the dev branch; running the same tests *after*
pushing is redundant and a waste of time.

For consistency, run codespell.yml on "merge_group" events too,
although I think this is unlikely to matter.