/etc/nginx: support PROXY protocol on ports 943 and 980.
We want to be able to put the nginx server behind a reverse proxy.
When doing so, we still would like to know the IP address of each
connecting client.

The "PROXY protocol" is a simple header that can be inserted at the
start of a TCP connection to inform the server of the real client IP
address.  haproxy can be configured to insert this header by setting
the server option 'send-proxy-v2'.

On the nginx side, the 'proxy_protocol' option must be set in the
'listen' line; this means that nginx will accept connections on that
port with the PROXY header (and will reject connections that don't
have the header.)

We use the nonstandard (but privileged) ports 980 and 943 for
HTTP-with-PROXY-header and HTTPS-with-PROXY-header respectively.
These ports will be blocked at the firewall level except for trusted
proxies.

With this done, we want nginx to treat the client address from the
header as the "real" IP address (recorded in logs and passed on to
uwsgi.)  We can do this unconditionally in the same server block,
since if the client isn't using the PROXY protocol, then
'real_ip_header proxy_protocol' does nothing.