get_associated_tasks: allow filtering by name.
In some cases, it's useful to be able to search for pending tasks
with a particular name as well as a particular associated object.

(The specified name must be a registered task name, i.e. it must be
the name of a function that is decorated by the associated_task
decorator.)