console: don't schedule a GCP/AWS upload if already pending.
In the published project management page, there are buttons to trigger
uploading project files to GCP and to AWS.  When the button is clicked
we want to schedule a background task that will do the actual work of
uploading.

If such a background task already exists (whether it's currently
running, or scheduled to run in the future) and the button is clicked
again, we don't want to schedule a second task.  In this case, simply
display a message and don't do anything.