manage_published_project: improve logic for cloud upload buttons.
In the published project management page, there are buttons to trigger
uploading project files to GCP and to AWS.  We want these buttons to
be hidden/disabled if the project is already scheduled to be uploaded,
so that administrators won't try to upload the project twice in a row.

Previously, the page assumed that if the GCP or AWS object existed but
its "sent_files" flag was false, then that meant there was an upload
pending.  This is incorrect in both directions:

- If the initial upload task failed, then "sent_files" would still be
  false.  Also, administrators could manually reset the "sent_files"
  flag if the cloud mirror was removed.  In either case, the button to
  re-upload should be enabled.

- If the initial upload task succeeded and then an administrator
  triggered a re-upload, then "sent_files" would still be true, but
  there would be a pending Task.  In that case, the button to
  re-upload should be disabled.

So instead of relying on the current state of the sent_files flag, we
should rely on whether a corresponding Task object exists.

The bucket name and status are now shown by
cloud_mirror_info_card.html, so we can remove the additional
explanatory text and simplify logic a little bit.